status/src/main/java/com/bootcamp/phincon/status/model/Notif.java
package com.bootcamp.phincon.status.model;


import jakarta.persistence.Entity;
import jakarta.persistence.Id;
import lombok.Data;
import org.springframework.data.relational.core.mapping.Table;

@Table(name="notifs")



@Entity

public class Notif {

    @Id
    String id;
    String status;

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getStatus() {
        return status;
    }

    public void setStatus(String status) {
        this.status = status;
    }
}
